intframes should be ints
the logic in the next lines may be incorrect, working on that next

+++ DataConverter.c
@@ -29,37 +29,13 @@ int UTF8toCodeP (uint8_t fileData)
     codepoint = fileData;
     printf("%08X\n",codepoint);
     return 0;
-
-  //two bytes
-//} //else if ((fileData >= 0xC0) && (fileData <= 0xDF)) {
-  //this should be able to handle multiple bytes
-  //  codepoint = ((fileData & 0x1F) << 6) | ((fileData) & 0x3F);
-  //  printf("%08X\n",codepoint);
-  //  return 0;
-
-  //three bytes
-//} //else if ((fileData >= 0xE0) && (fileData <= 0xEF)) {
-  //this should be able to handle multiple bytes
-  //  codepoint = (((fileData & 0x0F) << 12) | (((fileData) & 0x3F) << 6) |
-  //  ((fileData) & 0x3F));
-  //  printf("%08X\n",codepoint);
-  //  return 0;
-  //four bytes
-//} //else //((fileData >= 0xF0) && (fileData <= 0xF7)) {
-  //{
-    //this should be able to handle multiple bytes
-    //codepoint = (((fileData & 0x07) << 18) | ((fileData & 0x3F) << 12) |
-    //((fileData & 0x3F) <<  6) | ((fileData & 0x3F)));
-    //printf("%08X\n",codepoint);
-    //return 0;
-  //}
 }
 
 int twoFrames(uint8_t byte1,uint8_t byte2)
 {
   char char1[4];
   sprintf(char1,"%d%d",byte1,byte2);
-  uint8_t intframe1 = (uint8_t)strtol(char1,NULL,16);
+  int intframe1 = (int)strtol(char1,NULL,16);
   uint8_t codepoint1 = ((intframe1 & 0x1F) << 6) | ((intframe1) & 0x3F);
   printf("%08X\n",codepoint1);
   return 0;
@@ -69,7 +45,7 @@ int threeFrames(uint8_t byte1,uint8_t byte2,uint8_t byte3)
 {
   char char2[6];
   sprintf(char2,"%d%d%d",byte1,byte2,byte3);
-  uint8_t intframe2 = (uint8_t)strtol(char2,NULL,16);
+  int intframe2 = (int)strtol(char2,NULL,16);
   uint8_t codepoint2 = (((intframe2 & 0x0F) << 12) | (((intframe2) & 0x3F) << 6) |
   ((intframe2) & 0x3F));
   printf("%08X\n",codepoint2);
@@ -80,7 +56,7 @@ int fourFrames(uint8_t byte1,uint8_t byte2,uint8_t byte3, uint8_t byte4)
 {
   char char3[8];
   sprintf(char3,"%d%d%d%d",byte1,byte2,byte3,byte4);
-  uint8_t intframe3 = (uint8_t)strtol(char3,NULL,16);
+  int intframe3 = (int)strtol(char3,NULL,16);
   uint8_t codepoint3 = (((intframe3 & 0x07) << 18) | ((intframe3 & 0x3F) << 12) |
   ((intframe3 & 0x3F) <<  6) | ((intframe3 & 0x3F)));
   printf("%08X\n",codepoint3);
